+++ tests/react-router/vite.config.ts
@@ -1,0 +1,9 @@
+import { openimg } from "openimg/vite";
+import { reactRouter } from "@react-router/dev/vite";
+import tailwindcss from "@tailwindcss/vite";
+import { defineConfig } from "vite";
+import tsconfigPaths from "vite-tsconfig-paths";
+
+export default defineConfig({
+  plugins: [openimg(), tailwindcss(), reactRouter(), tsconfigPaths()],
+});