+++ packages/react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 0.1.2
+
+- Fix remove `background` styles once image is loaded.
+
 # 0.1.1
 
 - Expose `defaultGetSrc` function.